Fetches new orc8r_base_db and fixes minor issues
