+++ orc8r-lte-operator/metadata.yaml
@@ -28,7 +28,7 @@ provides:
     limit: 1
 
 requires:
-  db:
+  database:
     interface: postgresql_client
     limit: 1
