+++ orc8r-metricsd-operator/tests/integration/test_integration.py
@@ -20,6 +20,7 @@ SERVICE_REGISTRY_METADATA = yaml.safe_load(
 )
 
 APPLICATION_NAME = "orc8r-metricsd"
+DB_APPLICATION_NAME = "postgresql-k8s"
 CHARM_NAME = "magma-orc8r-metricsd"
 CERTIFIER_APPLICATION_NAME = "orc8r-certifier"
 CERTIFIER_CHARM_NAME = "magma-orc8r-certifier"
@@ -64,7 +65,7 @@ class TestOrc8rMetricsd:
 
     @staticmethod
     async def _deploy_postgresql(ops_test):
-        await ops_test.model.deploy("postgresql-k8s", application_name="postgresql-k8s")
+        await ops_test.model.deploy(DB_APPLICATION_NAME, application_name=DB_APPLICATION_NAME)
 
     @staticmethod
     async def _deploy_alertmanager(ops_test):
@@ -165,7 +166,7 @@ class TestOrc8rMetricsd:
             series="jammy",
         )
         await ops_test.model.add_relation(
-            relation1=CERTIFIER_APPLICATION_NAME, relation2="postgresql-k8s:db"
+            relation1=CERTIFIER_APPLICATION_NAME, relation2=f"{DB_APPLICATION_NAME}:db"
         )
         await ops_test.model.add_relation(
             relation1=CERTIFIER_APPLICATION_NAME, relation2="tls-certificates-operator"
@@ -197,7 +198,7 @@ class TestOrc8rMetricsd:
             series="jammy",
         )
         await ops_test.model.add_relation(
-            relation1=ACCESSD_APPLICATION_NAME, relation2="postgresql-k8s:db"
+            relation1=ACCESSD_APPLICATION_NAME, relation2=f"{DB_APPLICATION_NAME}:database"
         )
 
     async def _deploy_orc8r_orchestrator(self, ops_test):
