+++ orc8r-orchestrator-operator/tests/integration/test_integration.py
@@ -27,6 +27,7 @@ SERVICE_REGISTRY_APPLICATION_NAME = "orc8r-service-registry"
 SERVICE_REGISTRY_CHARM_NAME = "magma-orc8r-service-registry"
 SERVICE_REGISTRY_CHARM_FILE_NAME = "magma-orc8r-service-registry_ubuntu-22.04-amd64.charm"
 DOMAIN = "whatever.com"
+DB_APPLICATION_NAME = "postgresql-k8s"
 
 
 class TestOrchestrator:
@@ -50,7 +51,7 @@ class TestOrchestrator:
 
     @staticmethod
     async def _deploy_postgresql(ops_test):
-        await ops_test.model.deploy("postgresql-k8s", application_name="postgresql-k8s")
+        await ops_test.model.deploy(DB_APPLICATION_NAME, application_name=DB_APPLICATION_NAME)
 
     @staticmethod
     async def _deploy_tls_certificates_operator(ops_test):
@@ -111,7 +112,7 @@ class TestOrchestrator:
             series="jammy",
         )
         await ops_test.model.add_relation(
-            relation1=ACCESSD_APPLICATION_NAME, relation2="postgresql-k8s:db"
+            relation1=ACCESSD_APPLICATION_NAME, relation2=f"{DB_APPLICATION_NAME}:database"
         )
 
     async def _deploy_orc8r_certifier(self, ops_test):
@@ -134,7 +135,7 @@ class TestOrchestrator:
             series="jammy",
         )
         await ops_test.model.add_relation(
-            relation1=CERTIFIER_APPLICATION_NAME, relation2="postgresql-k8s:db"
+            relation1=CERTIFIER_APPLICATION_NAME, relation2=f"{DB_APPLICATION_NAME}:db"
         )
         await ops_test.model.add_relation(
             relation1=CERTIFIER_APPLICATION_NAME, relation2="tls-certificates-operator"
