+++ orc8r-policydb-operator/metadata.yaml
@@ -25,6 +25,6 @@ provides:
     limit: 1
 
 requires:
-  db:
+  database:
     interface: postgresql_client
     limit: 1
