+++ orc8r-subscriberdb-cache-operator/metadata.yaml
@@ -23,6 +23,6 @@ provides:
     limit: 1
 
 requires:
-  db:
+  database:
     interface: postgresql_client
     limit: 1
